+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog
 1.0a5 (unreleased)
 ------------------
 
+- When cookies are being set on the login view, return an html page that does a redirect in JavaScript.
+  When no cookies are being set, redirect to the OIDC server directly, as we always did until now.
+  This is done because mixing a redirect and a Set-Cookie header in one request may not work.
+  Fixes `issue 11 <https://github.com/collective/pas.plugins.oidc/issues/11>`_.
+  [maurits]
+
 - Update the plugin to make challenges.
   An anonymous user who visits a page for which you have to be authenticated,
   is redirected to the new require_login view on the plugin.